third_party/blink/web_tests/platform/generic/fast/forms/label/selection-disabled-label-expected.txt

Tests that the click event is passed when selection is disabaled on label

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getSelection().toString() is ""
PASS window.getSelection().toString() is "Select"
PASS document.querySelector('input').checked is false
PASS document.querySelector('input').checked is true
PASS document.querySelector('input').checked is false
PASS successfullyParsed is true

TEST COMPLETE
